Tackle https://github.com/django-import-export/django-import-export/pull/1695
