+++ stubs/django-import-export/import_export/utils.pyi
@@ -1,5 +1,7 @@
 from types import TracebackType
+from typing import Any, overload
 
+from django.db.models import Field as DjangoField, ForeignObjectRel, Model
 from django.db.transaction import Atomic
 
 class atomic_if_using_transaction:
@@ -10,3 +12,8 @@ class atomic_if_using_transaction:
     def __exit__(
         self, exc_type: type[BaseException] | None, exc_value: BaseException | None, exc_tb: TracebackType | None
     ) -> None: ...
+
+@overload
+def get_related_model(field: ForeignObjectRel) -> Model: ...
+@overload
+def get_related_model(field: DjangoField[Any, Any]) -> Model | None: ...
